cleanup of RoomInfoFragment
