+++ QLife/app/src/main/java/engsoc/qlife/utility/Constants.java
@@ -8,4 +8,5 @@ public class Constants {
     public static final String CALENDAR_FILE = "cal.ics";
     public static final String TV = "TV";
     public static final String PROJECTOR = "projector";
+    public static final String HTTP="http";
 }
