+++ QLife/app/src/main/res/layout/fragment_room_avaliability_info.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="engsoc.qlife.ui.fragments.RoomAvaliabilityInfoFragment">
+    tools:context="engsoc.qlife.ui.fragments.RoomInformationFragment">
 
     <android.support.v4.widget.NestedScrollView
     android:layout_width="match_parent"
